img2txt: fix an integer overflow in the BMP loader.
Fixes: #37 (CVE-2018-20545)
Fixes: #40 (CVE-2018-20548)
Fixes: #41 (CVE-2018-20549)

+++ src/common-image.h
@@ -1,19 +1,19 @@
 /*
  *  Imaging tools for cacaview and img2irc
- *  Copyright (c) 2003-2012 Sam Hocevar <sam@hocevar.net>
- *                All Rights Reserved
+ *  Copyright © 2003—2018 Sam Hocevar <sam@hocevar.net>
+ *              All Rights Reserved
  *
  *  This program is free software. It comes without any warranty, to
  *  the extent permitted by applicable law. You can redistribute it
  *  and/or modify it under the terms of the Do What the Fuck You Want
- *  to Public License, Version 2, as published by Sam Hocevar. See
- *  http://www.wtfpl.net/ for more details.
+ *  to Public License, Version 2, as published by the WTFPL Task Force.
+ *  See http://www.wtfpl.net/ for more details.
  */
 
 struct image
 {
     char *pixels;
-    unsigned int w, h;
+    size_t w, h;
     struct caca_dither *dither;
     void *priv;
 };
